Update Ubuntu 20 rabbitmq install - cloudsmith.io
I live in cuba, a USA forbidden country, so we (and other countries around the world) can't use the previously recommended site. So I used the official Installation guide and it works perfectly. Then I propose use it on the installations steps only, and keep the configuration as it is.
fix GeoNode/documentation#182